requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
pytz>=2023.3
PyYAML>=6.0
lxml>=4.9.2
pandas>=1.5.3
tenacity>=8.0.1
